+++ Source/dsp/DryLPF.cpp
@@ -19,7 +19,7 @@ void DryLPF::setCutoff(float val)
     //skew here: 
     // y = 0 to 1
     // y = ((x-min)/(max-min))^factor
-    float skewedValue = std::powf((val - 0.0f) / (10.0f - 0.0f), 0.1f);
+    float skewedValue = std::pow((val - 0.0f) / (10.0f - 0.0f), 0.1f);
 
     auto cutoffInHz = jmap(skewedValue, 0.0f, 1.0f, 20000.0f, 20.0f);
     cutoffSmoothed.setTargetValue(cutoffInHz);
